Refactor: split NotificationService into channels/quiet-hours/sound/dispatch + dedupe (#166)
NotificationService had grown to ~1,480 lines doing four unrelated jobs
(SRP): channel registry, quiet-hours/silent-mode policy, sound/vibration
playback, and notification building/dispatch, plus dead static-lifecycle
plumbing. All the prerequisite fixes this refactor was sequenced behind
(#153/#154/#155/#160/#165) are merged, so this is a pure move — no
behaviour change.

Split into focused, package-private collaborators in the service package:

- NotificationChannels — channel registry, versioned-ID scheme (#153),
  create/update/refresh, channelFor.
- QuietHours — the (mostly pure) quiet-hours/silent-mode policy:
  isWithinNotificationWindow / isWithinTimeRange / boundOrDefaultMinutes /
  alertsAllowedNow / shouldIgnoreSilentMode.
- AlertSounds — playAlarm + its soundExecutor + the DND check.
- OngoingStatusContent — the ongoing notification's title/detail/expanded
  text and small-icon selection (#192/#194).
- AlertSpec (now a shared top-level record) + NotificationConfig moved to
  their own files; NotificationService stays the public dispatch facade.

DRY items folded in:

1. One builder chain — the setTicker/…/BigTextStyle sequence written three
   times now lives in one alertBuilder(context, channelId, AlertSpec); the
   level, temperature, fast-drain, slow-charge and charge-connected alerts
   all flow through it.
2. One charge-style resolver — deliverPerChargeStyle(context, content,
   notificationDelivery) is shared by notifyChargeConnected and
   sendSlowChargeWarning (read pref → resolve → none/toast/notification).
3. One vibration pattern — VIBRATION_PATTERN now lives in SystemService and
   is referenced by channel creation, so channel vibration and the manual
   silent-mode buzz can't drift.
4. Dead lifecycle API removed — shutdown()/releaseCachedBitmap() had no
   callers and there is no Application class to wire them to; the executor
   and the WeakReference'd icon are reclaimed on process death.

Also consolidated the "Vibrate" preference read (previously an inline true
default in three spots) into AppPrefs.vibrateEnabled (the #162 facade).

Tests split to mirror the new classes: QuietHoursTest,
NotificationChannelsTest, OngoingStatusContentTest; NotificationServiceTest
trimmed to the charge-connected wiring + charge-style normalization; a
vibrateEnabled case added to AppPrefsTest. 446 tests + lintDebug +
assembleDebug green; installed on the Mate 10 Pro. No user-facing strings.

Closes #166
